+++ sdk/resources/azure-mgmt-resource/azure/mgmt/resource/templatespecs/v2022_02_01/aio/operations/__init__.py
@@ -6,8 +6,10 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 
-from ._operations import ChangesOperations
+from ._operations import TemplateSpecsOperations
+from ._operations import TemplateSpecVersionsOperations
 
 __all__ = [
-    'ChangesOperations',
+    'TemplateSpecsOperations',
+    'TemplateSpecVersionsOperations',
 ]